feat : 결제 검증 조회 기능 수정 및 결제 취소 로직 추가

+++ src/main/java/com/paymentapp/api/payment/dto/CreatePaymentRequest.java
@@ -1,7 +1,9 @@
 package com.paymentapp.api.payment.dto;
 
+import java.math.BigDecimal;
+
 public record CreatePaymentRequest(
         Long orderId,
-        Double totalAmount
+        BigDecimal totalAmount
 ) {
 }
